build(store): move ts to module/games/store
