+++ src/client/src/app/modules/games/module-games.module.ts
@@ -3,16 +3,19 @@ import { CommonModule } from '@angular/common';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { AddGameComponent } from './components/add-game/add-game.component';
 import { RoutingModule } from './routing.module';
-import { gameFeatureKey, reducer } from 'src/app/store/reducers/game/game.reducer';
+// import { gameFeatureKey, reducer } from 'src/app/modules/games/store/game.reducer';
+import * as fromGame from 'src/app/modules/games/store/game.reducer';
+
 import { EffectsModule } from '@ngrx/effects';
 import { StoreModule } from '@ngrx/store';
-import { GameEffects } from 'src/app/store/effects/game/game.effects';
+import { GameEffects } from 'src/app/modules/games/store/game.effects';
 
 
 
 @NgModule({
   declarations: [
     AddGameComponent,
+    
   ],
   imports: [
     CommonModule,
@@ -20,7 +23,7 @@ import { GameEffects } from 'src/app/store/effects/game/game.effects';
     ReactiveFormsModule,
     RoutingModule,
     EffectsModule.forFeature([GameEffects]),
-    StoreModule.forFeature(gameFeatureKey, reducer),
+    StoreModule.forFeature(fromGame.gameFeatureKey, fromGame.reducer),
   ]
 })
 export class GamesModule { }
