+++ src/client/src/app/services/game.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { map } from 'rxjs/operators';
 import { Game } from '../../../../shared/models/game.model';
-import { addGame } from '../store/actions/game/game.actions';
+import { addGame } from '../modules/games/store/game.actions';
 import { ApiService } from './api.service';
 
 @Injectable({
